requirements.txt
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
debugpy==1.8.0
firebase==4.0.1
firebase-admin==6.2.0
gcloud==0.17.0
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
jws==0.1.3
msgpack==1.0.7
numpy==1.26.1
oauth2client==3.0.0
opencv-python==4.8.1.78
picamera2==0.3.12
pidng==4.0.9
piexif==1.1.3
Pillow==10.1.0
proto-plus==1.22.3
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-jwt==2.0.1
python-prctl==1.8.1
requests==2.31.0
requests-toolbelt==0.7.0
rsa==4.9
simplejpeg==1.7.2
six==1.16.0
uritemplate==4.1.1
urllib3==2.0.7
v4l2-python3==0.3.2
